requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
yfinance>=0.2.0
plotly>=5.15.0
streamlit-option-menu>=0.3.0